move removal_evaluator
